% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_clusters.R
\name{assign.clusters}
\alias{assign.clusters}
\title{Assign.clusters}
\usage{
assign.clusters(data, cluster.list, cluster.pvalues, cluster.height,
  p.limit = 0.001, max.size = 1, max.tree.height = max(cluster.height))
}
\arguments{
\item{data}{}

\item{cluster.list}{}

\item{cluster.pvalues}{}

\item{cluster.height}{}

\item{p.limit}{}

\item{max.size}{}

\item{max.tree.height}{}
}
\value{
Return a data frame with an additional column to the data containing the cluster numbers
}
\description{
Assign.clusters
}
