% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_algo.R
\name{cluster.algo}
\alias{cluster.algo}
\title{Clustering algorithm using time, geographic space and genetic sequence information (Ypma et al 2013)}
\usage{
cluster.algo(data, time.col, x.col, y.col, gen.col, parallel = FALSE)
}
\arguments{
\item{data}{}

\item{time.col}{}

\item{x.col}{}

\item{y.col}{}

\item{gen.col}{}

\item{parallel}{}
}
\value{
clustering algortihm returns list with clusterids, clusterheights and pvalues
}
\description{
Clustering algorithm using time, geographic space and genetic sequence information (Ypma et al 2013)
}
\author{
J.Backer and L.Soetens
}
