% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_app.R
\name{cluster_app}
\alias{cluster_app}
\title{Cluster_app function}
\usage{
cluster_app(data, cluster.list, cluster.pvalue, cluster.height, id.col,
  time.col, x.col, y.col, gen.col, cols)
}
\arguments{
\item{data}{}

\item{cluster.list}{}

\item{cluster.pvalue}{}

\item{cluster.height}{}

\item{id.col}{}

\item{time.col}{}

\item{x.col}{}

\item{y.col}{}

\item{gen.col}{}

\item{cols}{}
}
\value{
An shiny application visualizing possilbe transmission clusters
}
\description{
Cluster_app function
}
