% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diss_relative.R
\name{diss.relative}
\alias{diss.relative}
\title{diss.relative}
\usage{
diss.relative(D)
}
\arguments{
\item{D}{}
}
\value{
estimates a relative distance matrix based on appendix S1 (Ypma et al. 2013)
}
\description{
diss.relative
}
